requirements.txt
numpy>=1.19.5
pandas>=1.3.0
torch>=1.9.0
torchvision>=0.10.0
scikit-learn>=1.0.0
matplotlib>=3.4.3
seaborn>=0.11.2
pyyaml>=6.0
nltk>=3.6.3
joblib>=1.1.0
tqdm>=4.62.0
